+++ medic-init/src/main.rs
@@ -1,6 +1,11 @@
+use medic_init::cli::CliArgs;
+use medic_init::create_config_file;
 use medic_src::AppResult;
 
+use clap::Parser;
+
 fn main() -> AppResult<()> {
-    println!("Hello, world!");
-    AppResult::Ok(())
+    let cli_args = CliArgs::parse();
+
+    create_config_file(cli_args.config)
 }