+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Add `medic init` to generate an *almost* empty manifest file.
 - Fix `--config` description in medic help.
 
 #### Breaking Changes: